Pull the OAI parser into its own file
